fix: ajuste de las respuestas del crud

+++ frankenstein/src/main/java/com/endorodrigo/frankenstein/controller/CustomerController.java
@@ -1,56 +1,64 @@
 package com.endorodrigo.frankenstein.controller;
 
 import com.endorodrigo.frankenstein.entity.Customer;
-import com.endorodrigo.frankenstein.repository.CustomerRepository;
 import com.endorodrigo.frankenstein.servicies.CustomerService;
+import java.util.List;
+import java.util.Optional;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/home")
+@RequestMapping("/api/customer")
 @Slf4j
 public class CustomerController {
 
-    private CustomerService customerService;
+    private final CustomerService customerService;
 
     public CustomerController(CustomerService customerService) {
         this.customerService = customerService;
     }
-    
 
     // Mostrar todos los clientes
     @GetMapping
-    public String listCustomers(Model model) {
-        model.addAttribute("customers", customerService.findAll());
-        model.addAttribute("customer", new Customer());
-        return "home";
+    public ResponseEntity<List<Customer>> listCustomers() {
+        List<Customer> customers = customerService.findAll();
+        return new ResponseEntity<>(customers, HttpStatus.OK);
     }
 
     // Guardar un nuevo cliente o actualizar uno existente
     @PostMapping
-    public String saveCustomer(@ModelAttribute Customer customer) {
-        customerService.save(customer);
-        log.info("=== creando customer ==");
-        log.info(customer.toString());
-        return "redirect:/home";
+    public ResponseEntity<Customer> saveCustomer(@RequestBody Customer customer) {
+        Customer savedCustomer = customerService.save(customer);
+        return new ResponseEntity<>(savedCustomer, HttpStatus.CREATED);
     }
 
-    // Editar un cliente
-    @GetMapping("/edit/{id}")
-    public String editCustomer(@PathVariable("id") Integer id, Model model) {
-        Customer customer = customerService.findById(id).orElseThrow();
-        model.addAttribute("customer", customer);
-        model.addAttribute("customers", customerService.findAll());
-        return "customerForm";
+    @PutMapping
+    public ResponseEntity<Customer> editCustomer(@RequestBody Customer data) {
+        Optional<Customer> findUser = customerService.findById(data.getId());
+        if (findUser.isPresent()) {
+            findUser.get().setName(data.getName());
+            findUser.get().setLastName(data.getLastName());
+            findUser.get().setCellNumber(data.getCellNumber());
+            findUser.get().setAddress(data.getAddress());
+            findUser.get().setStatus(data.isStatus());
+            return ResponseEntity.status(HttpStatus.CREATED).body(customerService.save(findUser.get()));
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     // Eliminar un cliente
-    @GetMapping("/delete/{id}")
-    public String deleteCustomer(@PathVariable("id") Integer id) {
-        customerService.deleteById(id);
-        return "redirect:/customers";
+    @DeleteMapping
+    public ResponseEntity<Void> deleteCustomer(@PathVariable("id") Integer id) {
+        Optional<Customer> customer = customerService.findById(id);
+        if (customer.isPresent()) {
+            customerService.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
+
 }
